move scripts to github actions
