+++ .gitignore
@@ -6,6 +6,7 @@ node_modules
 
 # Build directory
 /public
+/og-images
 .DS_Store
 
 # Bundle / eager-graph check artifacts (generated by the EMIT_WEBPACK_STATS build)
